# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "receipt-system",
    "description": "AI\u99c6\u52d5\u578b\u78ba\u5b9a\u7533\u544a\u7528\u30ec\u30b7\u30fc\u30c8\u7ba1\u7406\u30b7\u30b9\u30c6\u30e0 - Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "\u8a2d\u5b9a\u4e00\u89a7\u53d6\u5f97",
        "description": "\u30b7\u30b9\u30c6\u30e0\u306e\u8a2d\u5b9a\u4e00\u89a7\u3092\u53d6\u5f97\u3057\u307e\u3059\u3002",
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsListResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "\u8a2d\u5b9a\u4e00\u62ec\u66f4\u65b0",
        "description": "\u8907\u6570\u306e\u8a2d\u5b9a\u3092\u4e00\u62ec\u3067\u66f4\u65b0\u3057\u307e\u3059\u3002",
        "operationId": "update_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateMultipleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/{key}": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "\u8a2d\u5b9a\u66f4\u65b0",
        "description": "\u6307\u5b9a\u3055\u308c\u305f\u8a2d\u5b9a\u3092\u66f4\u65b0\u3057\u307e\u3059\u3002",
        "operationId": "update_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SettingsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SettingsResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "SettingsListResponse",
        "description": "Settings list response schema."
      },
      "SettingsResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "value"
        ],
        "title": "SettingsResponse",
        "description": "Settings response schema."
      },
      "SettingsUpdateMultipleRequest": {
        "properties": {
          "settings": {
            "items": {
              "$ref": "#/components/schemas/SettingsUpdateRequest"
            },
            "type": "array",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "settings"
        ],
        "title": "SettingsUpdateMultipleRequest",
        "description": "Settings multiple update request schema."
      },
      "SettingsUpdateRequest": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "SettingsUpdateRequest",
        "description": "Settings update request schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}